Move lunatik.h to include/linux/
Move lunatik.h from lib/lunatik/ to include/linux/ for nicer include
paths.  To not have in turn a very relative include of lauxlib.h in
lunatik.h add a proxy include file for it in a new subdirectory
include/linux/lunatik/.  It is very probable that other Lua includes may
follow this practice, when more bindings get added.

The original include files within the lua subdir of lunatik shall not
move, because that would incur changes to the original lua sources,
which changes we want to minimize.  So proxying them seems to be the
best for both sides (better integration AND minimal changes).