Add variant of loadcode not using a workqueue
Add lunatik_loadcode_direct which is calling the Lua interpreter in the
current context.  The caller needs to ensure that the context is an
appropriate process context.  To ensure that the workqueue driven
interface is not interfering with that new interface put a mutex around
the actual Lua state touching part.