Introduce context abstraction in Lunatik
A Lunatik context is a wrapper around a Lua state.  Each context
maintains its own workqueue but also can be used in direct mode if we
are already in process context.  A Lunatik context also carries a mutex
that shall be locked if the internal Lua state is used outside of the
Lunatik functions which include the locking.

Remove separate lunatik_workqueue.c/h as the workqueue is exposed
through struct lunatik_context and therefore needed to be declared in
lunatik.h.  The implementations are also very short leaving minimal
argument for a separate implementation file.

The bindings (currently) enrich a default context which is also the
context used for the Lua syscall.  This is about to change and the
default context is going to be removed in the future.  Instead it is
planned to have a dedicated syscall context and an infrastructure where
bindings can register their install functions.

At the end this will allow syscall and other users to be isolated from
each other Lua wise.