Introduce registry for Lunatik bindings
Every binding registers a register function.  With a call of another
function all registered bindings can be loaded into a Lunatik context.
This allows the bindings being used in other Lunatik contexts without
knowing them explicitely.  That improves the extendability of the
bindings concept.

The default context is repeatedly populated with the bindings on every
get to ensure that new bindings are available in that context (e.g. for
Lua syscall etc.).